update int to long

+++ StockService.java
@@ -7,7 +7,7 @@ public class StockService {
     private static final int threshold=3;
 
     // Προσθήκη αντικειμένου τύπου stock στην λίστα με όλα τα αποθέματα
-    public void addStock(int storeId, int productId, int stockQuantity) { 
+    public void addStock(long storeId, long productId, int stockQuantity) { 
     	if (storeId <= 0 || productId <= 0) {
             System.out.println("Μη έγκυρο κατάστημα ή προϊόν.");
             return;
@@ -23,7 +23,7 @@ public class StockService {
     
 
     //Έλεγχος αποθέματος για συγκεκριμένο προιόν και κατάστημα 
-    public List<Stock> getStock(int productId, int storeId) {   
+    public List<Stock> getStock(long productId, long storeId) {   
         List<Stock> result = new ArrayList<>();
 
         if (stocks != null && !stocks.isEmpty()) { 
@@ -39,7 +39,7 @@ public class StockService {
 
     
     // Ενημέρωση αποθέματος 
-    public void updateStock(int productId, int storeId, int newQuantity) {  
+    public void updateStock(long productId, long storeId, int newQuantity) {  
         List<Stock>currentStocks=getStock(productId, storeId);
         if (currentStocks!=null){
             for(Stock s : currentStocks){
@@ -50,7 +50,7 @@ public class StockService {
 
     
     //Μειώνει το απόθεμα
-    public void reduceStockOnPurchase(int productId, int storeId, int quantity) { 
+    public void reduceStockOnPurchase(long productId, long storeId, int quantity) { 
         List<Stock> stockList = getStock(productId, storeId);
         if (!stockList.isEmpty()) {
             // Παίρνουμε το πρώτο διαθέσιμο απόθεμα για το προϊόν και το κατάστημα
@@ -79,7 +79,7 @@ public class StockService {
     }
     
     // Αναζητά εάν υπάρχει το προιόν σε άλλο κατάστημα διαθέσιμο
-    public ArrayList<Stock> searchProductInOtherStores(int productId, int excludedStoreId) {
+    public ArrayList<Stock> searchProductInOtherStores(long productId, long excludedStoreId) {
         ArrayList<Stock> result = new ArrayList<>();  // Νέα λίστα για τα αποτελέσματα
     
         for (Stock s : stocks) {  
@@ -101,7 +101,7 @@ public class StockService {
     }
     
     //Θα επιστρέψει το απόθεμα ενός προϊόντος σε JSON μορφή, για όλα τα καταστήματα που το έχουν διαθέσιμο
-    public String getStockAsJson(int productId) {  
+    public String getStockAsJson(long productId) {  
         StringBuilder json = new StringBuilder();
         json.append("{\n  \"productId\": ").append(productId).append(",\n  \"stockPerStore\": [\n");
